Refactor: LinkSubFolder의 컴포넌트들을 Folder.js에서 재조립
모달 기능 추가시 나타날 것으로 예상되는 프롭 드릴링 현상을 최소화함
